requirements.txt
Pillow
pytesseract
Flask
requests 
opencv-python
werkzeug
